requirements.txt
pyyaml>=6.0
pandas>=2.0.3
requests>=2.31.0
numpy>=1.24.4
tqdm>=4.65.0
ruff>=0.0.291